Move styling + adjust diff styling
