+++ nbdime/webapp/src/app/common.ts
@@ -2,7 +2,7 @@
 // Distributed under the terms of the Modified BSD License.
 'use strict';
 
-import '../common/cm-merge.css';
+import './cm-merge.css';
 
 
 /**
