+++ nbdime/webapp/src/app/diff.css
@@ -16,6 +16,14 @@ body {
   width: 80%;
 }
 
+.jp-Diff-unchanged .jp-Output *:first-child {
+  margin-left: auto;
+}
+
+.jp-Diff-unchanged .jp-Output *:last-child {
+  margin-right: auto;
+}
+
 .jp-Diff-unchanged .jp-Diff-unchanged {
   margin-left: 0;
   width: 100%;
