+++ nbdime/webapp/src/app/diff.ts
@@ -2,7 +2,7 @@
 // Distributed under the terms of the Modified BSD License.
 'use strict';
 
-import '../diff/diff.css';
+import './diff.css';
 
 
 import {
